requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-extensions==2.2.1
django-filter==2.2.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
flake8-isort==2.7.0
idna==2.8
isort==4.3.21
mccabe==0.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
testfixtures==6.10.0
urllib3==1.25.6
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-extensions==2.2.1
django-filter==2.2.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
flake8-isort==2.7.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
mccabe==0.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
testfixtures==6.10.0
urllib3==1.25.6
